add migrate department_id to user, policy:departmant, user departmant, project, user project

+++ app/controllers/departments_controller.rb
@@ -5,6 +5,7 @@ class DepartmentsController < ApplicationController
 
   def index
     @departments = Department.paginate(page: params[:page], per_page: 10)
+    authorize @departments
   end
 
   def show
@@ -12,14 +13,17 @@ class DepartmentsController < ApplicationController
     @users = User.where(department_id: params[:id])
     @user_leader = User.find_by(id: @department.user_id)
     @project = Project.where(department_id: params[:id])
+    authorize @department
   end
 
   def new
     @department = Department.new
+    authorize @department
   end
 
   def create
     @department = Department.new(departments_param)
+    authorize @department
     if @department.save
       flash[:success] = "Create to Success!"
       redirect_to departments_path
@@ -32,10 +36,12 @@ class DepartmentsController < ApplicationController
 
   def edit
     @department = Department.find(params[:id])
+    authorize @department
   end
 
   def update
     @department = Department.find(params[:id])
+    authorize @department
     if @department.update(departments_param)
       redirect_to action: :index
       flash[:success] = "Update to Success!"
@@ -47,6 +53,7 @@ class DepartmentsController < ApplicationController
 
   def destroy
     @department = Department.find(params[:id])
+    authorize @department
     if @department.destroy
       flash[:success] = 'Department was successfully deleted.'
     else
@@ -54,7 +61,6 @@ class DepartmentsController < ApplicationController
     end
     redirect_to departments_path
   end
-
 
   def remove_user_from_department
     department = Department.find(department_id)
